Fix R CMD check note about undefined global variables.
https://community.rstudio.com/t/how-to-solve-no-visible-binding-for-global-variable-note/28887
https://cran.r-project.org/web/packages/dplyr/vignettes/programming.html

For most, used the rlang data pronoun ".data". As far as I can tell
there is no way to refer to the entire data frame (e.g. with "."), so I
had to add "." with utils::globalVariables().

https://rlang.r-lib.org/reference/tidyeval-data.html

Note that R CMD check does not complain about dplyr::filter(). However,
I still added .data$ for 1) consistency, and 2) to reduce ambiguity, as
demonstrated in the dplyr programming vignette.