Require RSQLite >= 2.1.1
I only define some of the columns with field.types when writing database
tables with DBI::dbWriteTable(). Prior to RSQLite 2.1.1, the columns
defined with field.types were assumed to be the leftmost columns in the
data frame. Since they aren't in my case, the columns were incorrectly
named. Thus when the terms were queried to perform the overlaps, none of
the termID's matched the terms in the enrichments table, because neither
table had the correct values in the column termID.

I discovered this bug when testing in rocker/tidyverse:3.4.3. After
updating tidyr and dbplyr, RSQLite was still at 2.0.

Here's the relevant NEWS bullet:

Breaking change: The field.types argument to dbWriteTable() no longer
takes precedence when defining the order of the columns in the new
table.

https://github.com/r-dbi/RSQLite/blob/master/NEWS.md#rsqlite-211-2018-05-05