Overhaul storage of enrichments results and refactor.
Changed from:

models -> tests -> annotations

to:

models -> annotations -> tests

This greatly simplifies the getX() functions.

The only downside is the intersection of the annotation terms and
enrichments results in getEnrichmentNetwork() is less efficient since
the full tables are returned instead of only the intersection. I think
this is a small price to pay for maintainability.