Overhaul plotting so that it works with study packages.
plotStudy() is now an S3 method that can also work on package names.

getPlots() is now identical for oaStudy and study package.

dynGet() was added in R 3.2.0
https://cran.r-project.org/doc/manuals/NEWS.3

The key was not passing the function to addPlots(), but instead passing
the name. The function definition is then fetched from the environment.
This complicates exporting to rds, but this isn't a critical feature.