Wrap EnrichmentUpsetPlot() and test getEnrichmentsUpset().
Had to debug EnrichmentUpsetPlot(). The test enrichments table had
an NA, which broke sum(). The NA occurs because it wasn't significant
across all the tests, and thus it wasn't included in the enrichment
results. The instructions encourage providing all enrichment results,
not just those that are significant, so this should be a rare occurence
in practice.