Teardown plot tests.
This ended up being a ton of work to figure out. My mistake of not
cleaning up after the plot tests was actually helping the tests. They
are run alphabetically, and the only tests after that install packages
is testUpset.R. Once I cleaned up, these tests started having strange
warnings about missing DESCRIPTION files. It took way too long to figure
out because I wasn't expecting what caused the problem. When
find.package() searches for a package, it looks in order of .libPaths().
Thus I assumed there would be no issue. But it wasn't finding the newly
installed OAstudyABC. Instead it was finding the deleted version from
testPlot.R. It turns out that find.package() first searches the packages
with loaded namespaces. Since the namespace was loaded to create the
plots, it always returned the path to this now deleted package. Thus
properly cleaning up required manually unloading the package namespace.

Here's the exact line in find.package() that caused the behavior I
wasn't anticipating:

https://github.com/wch/r-source/blob/157225dfb95ea6eee615f2c34d677a0abf2015f8/src/library/base/R/library.R#L792

.getNamespaceInfo(asNamespace(pkg), "path")

This is documented. The namespaces are searched prior to the libary
paths.