Improve messaging when exporting study to package.
When I overhauled this to plaintext, I didn't retain much messaging.
This was partly because the recursive function exportElements() made it
difficult to report progress.

I suppressed the potential messages from getPlots() and note when the
overlaps are being calculated, since that is the part that will take the
longest.