Be more explicit with I/O options.
For fread() - Only OmicAnalyzer reads/writes the files, so they will
always have a header and be tab-separated. This doesn't change the
behavior, but it avoids fread() haveing to guess. Hopefully this will
improve the speed and also avoid any potential edge cases.

For fwrite() - Explicitly surround columns that should be characters
with quotes. Unfortunately this is more for documentation, i.e. for
anyone that visually inspects the raw files. fread() will still convert
quoted numeric columns to be numbers, which is apparently part of the
CSV standard specs. Thus I may or may not keep this change, since all of
those quotation marks will cause the files to be bigger.

https://github.com/Rdatatable/data.table/issues/2052#issuecomment-289721855