Preserve featureID column as character vector.
The previous commit caused a regression. Currently the featureID column
in the results table isn't required to be a character. Longer term it'd
probably be better to expand coerceColsToCharacter() to apply to
specifc columns, and enforce this at the level of addResults().
Frustratingly this would still require some patch code to getResults()
because a numeric featureID would still get read in as numeric. And this
could be a problem for ID columns in other tables. Having a hard time
thinking of a quick fix that would apply to all tables.