Install study pkg directory from source directory.
I was building a tarball and installing from that. This step is fragile
because R CMD build creates the tarball in the current working
directory. Also, if R CMD build fails due to something wrong with the
package (e.g. invalid package name), this is hard to detect and report
well to the end user.