Formatting tweaks to User's Guide.
To fix the numbering of the references section in the TOC, I switched
to using an explicit \section{} before bibliography. This was inspired
by the thread below, though I put the \section{} outside of the
bibliography environment, which I think looks better.

https://groups.google.com/forum/#!topic/latexusersgroup/3_Q_3ReOrWA

The main downside is that the section title References is duplicated.