Don't include PDF diagrams in package tarball.
The final vignettes saved in inst/doc/ are built fine without including
the image files in the tarball. Users won't be able to build the
vignette from this package tarball, but they should download the source
repository if they want to do that.

I did this to save on space as much as possible, since the web app is
already so big.

$ du -sh vignettes/images/
232K    vignettes/images/