Avoid complexity of data.table copy by reference with setDT().
In order to modify the object without making a copy, it modifies the
object in the parent environment. In this case it was enrichmentsTable,
an intermediate object that is never used again. So it doesn't matter.
But when I tried to use setDF(x) to convert enrichmentsTable back to a
data frame, it didn't work. Thus I don't think it's worth trying to
manage these side effects unless there is a clear performance benefit.