Improve the validation for results table linkouts.
The initial version was too rigid. It now accepts the following
situations:

* There is no features table. Then it is still ok to have linkouts that
reference the featureID column, since this is still present in the
results table.

* If the linkouts apply to the special modelID "default", but the
features tables themselves are all specific to a separate modelID. This
is still fine as long as these tables still all have the necessary
columns.