Overhaul the validation functions to focus on app behavior.
The previous validations were too strict and also fragile (especially
when the modelID "default" was used). The new validation functions focus
on the app's behavior. For example, if there are no custom plotting
functions for a given modelID, then it's samples and assays tables
aren't checked, since it is irrelevant. And in general, they now throw
warnings instead of errors in the situation where the user may want to
add some more information.

I got the tests passing locally, but I need to overhaul these as well.
Especially to handle all the new warnings.