Send message instead of warning for validation alerts.
There is nothing wrong with the R code. These are just notifications
that certain features of the app will be disabled unless more
information is provided.

Also, only send message once for when the entire annotations is missing.
Before it was sending a message for each missing annotationID in the
enrichments table, for each modelID!