Support numeric-looking featureIDs in assays table.
Since the numeric featureIDs in my tests have leading zeros, the fread()
argument keepLeadingZeros was sufficient to fix the problem. But since
the featureIDs could be numeric-looking and not have any leading zeros,
I also explicitly set the first column to be character. Since
hasRowNames = TRUE, this initial first column gets moved to the row
names.