Sort the data frames returned by getPlottingData().
There is no requirement when adding the data frames to the OmicNavigator
study that the features and samples data frames match the order of the
rows and columns of the assays data frame, respectively. This is
essential for plotting, so getPlottingData() ensures that the rows of
the features data frame match the rows of the assays data frame, and
that the rows of the samples data frame match the columns of the assays
data frame. In this sense, it is a psuedo-ExpressionSet-like object.