Don't automatically download the app files during installation.
We decided this is more trouble than it was worth. `R CMD check`
flags large directories in the installed package, not just the source
code tarball. Also, this prevents a minimal installation of the R
package. Lastly, since `configure` has to run prior to installation,
it requires maintaining a second set of functions that can't assume
any of the internal package information.