Suport tibble and data.table input.
The checkX() fucntions now select columns with [[, so that a tibble
input also returns a vector (using [ returns a 1-column tibble).

The tibble or data.table input is converted to a data frame. This is
for internal consistency. I attempted to support internal usage of
tibbles and data tables, but this was tough. It broke a lot of stuff.

In general, the new sanitizeX() functions sanitize the input for
storage in the onStudy object.