Make the package data.table aware.
This was causing my earlier issues with "object not found" errors. The
functions worked, but not inside of a package because of namespace
issues.

https://rdatatable.gitlab.io/data.table/articles/datatable-importing.html#data-table-in-imports-but-nothing-imported-1
https://jozef.io/r911-datatable-cedta/

This also forced me to update the filtering in getResultsUpset(). It
relied on the behavhior of [.data.frame to convert a 1-column df to a
vector. [.data.table does not support this behavior.

https://rdatatable.gitlab.io/data.table/articles/datatable-faq.html#j-num

Hopefully this will also make those filtering operations faster now that
it is actually using data.table as I had intended.