Don't trigger deployment from Pull Requests.
https://docs.github.com/en/actions/reference/workflow-syntax-for-github-actions#example-using-contexts
https://docs.github.com/en/actions/reference/context-and-expression-syntax-for-github-actions#operators

Note: This would never deploy the code from the PR. If someone were to
edit the workflow file such that the deployment was triggered from a PR,
it would simply re-deploy the code that was already in the main branch.
In other words, the code that handles the actual deployment is in an
internal private repo unaccessible to the public. This simply triggers
that mechanism.