Remove study package from loaded namespace after import
This was causing difficult-to-debug errors in the unit tests.
Essentially, even though the package directory was deleted and the
.libPaths() restored, find.package() was still finding the deleted
package. This is because find.package() first searches the loaded
namespaces before searching .libPaths(). Thus it would find the
deleted package first, and then throw a warning, which caused the
plotting tests for silence to fail.

https://github.com/abbvie-external/OmicNavigator/runs/2915754869#step:8:964