Remove long-running example for getInstalledStudies()
Presumably it takes a lot longer to load a list of all the installed
packages on a CRAN machine with 10k+ packages. Note that predictably
this only affected the Windows machine and not Debian.

I don't know if this would have prevented the CRAN submission. There
was also a WARNING because leftover tex files broke the vignette build.
But I don't want to take any chances.