Ignore leftover intermediate LaTeX files
Best case these generate a NOTE. Worse case they can break the
vignette build and generate a WARNING (which is what happened on
the Windows pretest).

Only .bbl and .blg actually made it through. The rest must be
ignored by default. I added them to be explicit (and in case this
automatic behavior changes in the future). Lastly I didn't ignore
pdf files since these are already ignored, and this causes the
pdf images to be ignored.