Return plotting data invisibly. Explicitly print ggplot2 plots
This is for the user experience in the R console. When they run
plotStudy() with a ggplot2 plot, they will want it displayed
immediately. When they run with a base R plot, they don't want
the R console overwhelmed with the underlying data. These changes
should not impact the app.