Don't include manual dispatch for release job
This wouldn't work as is. It needs a tag to extract the version.
I would need to include dispatch inputs for it to work, but that
would really turn it into a project. Then the steps would have to
be conditional on how it was triggered. And it could get weird
since this is supposed to be tied to a tag.

https://docs.github.com/en/actions/learn-github-actions/workflow-syntax-for-github-actions#onworkflow_dispatchinputs