Support adding a single serialized R object per modelID (#90)
* Support adding a single serialized R object per modelID

* Test support for addObjects()

* Distinguish between a list and an object that inherits from the list class

This is required to handle the following note from `R CMD check`:

N  checking R code for possible problems (5.8s)
   Found if() conditions comparing class() to string:
   File 'OmicNavigator/R/check.R': if (class(objects[[i]]) == "list") ...
   Use inherits() (or maybe is()) instead.

I get why they added this note. It is a best practice. But in this
case, we are truly trying to distinguish between an object whose first
class is "list" and an object that may inherit from "list" later in its
class hierarchy.

* Document support for addObjects()